python-wrap-clouddrive-web-api/requirements.txt
grpcio
grpcio-tools
protobuf
requests
python-dateutil
yarl
